+++ mod-source-record-storage-server/src/main/java/org/folio/dao/SnapshotDaoImpl.java
@@ -1,7 +1,6 @@
 package org.folio.dao;
 
 import io.vertx.core.Future;
-import io.vertx.core.Vertx;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.sql.UpdateResult;
@@ -9,15 +8,17 @@ import org.folio.rest.jaxrs.model.Snapshot;
 import org.folio.rest.jaxrs.model.SnapshotCollection;
 import org.folio.rest.persist.Criteria.Criteria;
 import org.folio.rest.persist.Criteria.Criterion;
-import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.persist.cql.CQLWrapper;
 import org.folio.rest.persist.interfaces.Results;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
 import static org.folio.dataimport.util.DaoUtil.constructCriteria;
 import static org.folio.dataimport.util.DaoUtil.getCQLWrapper;
 
+@Component
 public class SnapshotDaoImpl implements SnapshotDao {
 
   private static final Logger LOG = LoggerFactory.getLogger(SnapshotDaoImpl.class);
@@ -25,19 +26,16 @@ public class SnapshotDaoImpl implements SnapshotDao {
   private static final String SNAPSHOTS_TABLE = "snapshots";
   private static final String SNAPSHOT_ID_FIELD = "'jobExecutionId'";
 
-  private PostgresClient pgClient;
-
-  public SnapshotDaoImpl(Vertx vertx, String tenantId) {
-    pgClient = PostgresClient.getInstance(vertx, tenantId);
-  }
+  @Autowired
+  private PostgresClientFactory pgClientFactory;
 
   @Override
-  public Future<SnapshotCollection> getSnapshots(String query, int offset, int limit) {
+  public Future<SnapshotCollection> getSnapshots(String query, int offset, int limit, String tenantId) {
     Future<Results<Snapshot>> future = Future.future();
     try {
       String[] fieldList = {"*"};
       CQLWrapper cql = getCQLWrapper(SNAPSHOTS_TABLE, query, limit, offset);
-      pgClient.get(SNAPSHOTS_TABLE, Snapshot.class, fieldList, cql, true, false, future.completer());
+      pgClientFactory.createInstance(tenantId).get(SNAPSHOTS_TABLE, Snapshot.class, fieldList, cql, true, false, future.completer());
     } catch (Exception e) {
       LOG.error("Error while querying snapshots", e);
       future.fail(e);
@@ -48,11 +46,11 @@ public class SnapshotDaoImpl implements SnapshotDao {
   }
 
   @Override
-  public Future<Optional<Snapshot>> getSnapshotById(String id) {
+  public Future<Optional<Snapshot>> getSnapshotById(String id, String tenantId) {
     Future<Results<Snapshot>> future = Future.future();
     try {
       Criteria idCrit = constructCriteria(SNAPSHOT_ID_FIELD, id);
-      pgClient.get(SNAPSHOTS_TABLE, Snapshot.class, new Criterion(idCrit), true, false, future.completer());
+      pgClientFactory.createInstance(tenantId).get(SNAPSHOTS_TABLE, Snapshot.class, new Criterion(idCrit), true, false, future.completer());
     } catch (Exception e) {
       LOG.error("Error querying snapshots by id", e);
       future.fail(e);
@@ -63,18 +61,18 @@ public class SnapshotDaoImpl implements SnapshotDao {
   }
 
   @Override
-  public Future<String> saveSnapshot(Snapshot snapshot) {
+  public Future<String> saveSnapshot(Snapshot snapshot, String tenantId) {
     Future<String> future = Future.future();
-    pgClient.save(SNAPSHOTS_TABLE, snapshot.getJobExecutionId(), snapshot, future.completer());
+    pgClientFactory.createInstance(tenantId).save(SNAPSHOTS_TABLE, snapshot.getJobExecutionId(), snapshot, future.completer());
     return future;
   }
 
   @Override
-  public Future<Boolean> updateSnapshot(Snapshot snapshot) {
+  public Future<Boolean> updateSnapshot(Snapshot snapshot, String tenantId) {
     Future<UpdateResult> future = Future.future();
     try {
       Criteria idCrit = constructCriteria(SNAPSHOT_ID_FIELD, snapshot.getJobExecutionId());
-      pgClient.update(SNAPSHOTS_TABLE, snapshot, new Criterion(idCrit), true, future.completer());
+      pgClientFactory.createInstance(tenantId).update(SNAPSHOTS_TABLE, snapshot, new Criterion(idCrit), true, future.completer());
     } catch (Exception e) {
       LOG.error("Error updating snapshots", e);
       future.fail(e);
@@ -83,9 +81,9 @@ public class SnapshotDaoImpl implements SnapshotDao {
   }
 
   @Override
-  public Future<Boolean> deleteSnapshot(String id) {
+  public Future<Boolean> deleteSnapshot(String id, String tenantId) {
     Future<UpdateResult> future = Future.future();
-    pgClient.delete(SNAPSHOTS_TABLE, id, future.completer());
+    pgClientFactory.createInstance(tenantId).delete(SNAPSHOTS_TABLE, id, future.completer());
     return future.map(updateResult -> updateResult.getUpdated() == 1);
   }