+++ mod-source-record-storage-server/src/main/java/org/folio/rest/impl/InitAPIImpl.java
@@ -1,0 +1,29 @@
+package org.folio.rest.impl;
+
+import io.vertx.core.AsyncResult;
+import io.vertx.core.Context;
+import io.vertx.core.Future;
+import io.vertx.core.Handler;
+import io.vertx.core.Vertx;
+import org.folio.config.ApplicationConfig;
+import org.folio.rest.resource.interfaces.InitAPI;
+import org.folio.spring.SpringContextUtil;
+
+public class InitAPIImpl implements InitAPI {
+
+  @Override
+  public void init(Vertx vertx, Context context, Handler<AsyncResult<Boolean>> handler) {
+    vertx.executeBlocking(
+      future -> {
+        SpringContextUtil.init(vertx, context, ApplicationConfig.class);
+        future.complete();
+      },
+      result -> {
+        if (result.succeeded()) {
+          handler.handle(Future.succeededFuture(true));
+        } else {
+          handler.handle(Future.failedFuture(result.cause()));
+        }
+      });
+  }
+}