+++ mod-source-record-storage-server/src/main/java/org/folio/services/RecordService.java
@@ -18,41 +18,46 @@ public interface RecordService {
    * @param query  query from URL
    * @param offset starting index in a list of results
    * @param limit  limit of records for pagination
+   * @param tenantId tenant id
    * @return future with {@link RecordCollection}
    */
-  Future<RecordCollection> getRecords(String query, int offset, int limit);
+  Future<RecordCollection> getRecords(String query, int offset, int limit, String tenantId);
 
   /**
    * Searches for record by id
    *
    * @param id Record id
+   * @param tenantId tenant id
    * @return future with optional {@link Record}
    */
-  Future<Optional<Record>> getRecordById(String id);
+  Future<Optional<Record>> getRecordById(String id, String tenantId);
 
   /**
    * Saves record
    *
    * @param record Record to save
+   * @param tenantId tenant id
    * @return future with true if succeeded
    */
-  Future<Boolean> saveRecord(Record record);
+  Future<Boolean> saveRecord(Record record, String tenantId);
 
   /**
    * Updates record with given id
    *
    * @param record Record to update
+   * @param tenantId tenant id
    * @return future with true if succeeded
    */
-  Future<Boolean> updateRecord(Record record);
+  Future<Boolean> updateRecord(Record record, String tenantId);
 
   /**
    * Deletes record by id
    *
    * @param id record id
+   * @param tenantId tenant id
    * @return future with true if succeeded
    */
-  Future<Boolean> deleteRecord(String id);
+  Future<Boolean> deleteRecord(String id, String tenantId);
 
   /**
    * Searches for source records
@@ -60,7 +65,8 @@ public interface RecordService {
    * @param query  query from URL
    * @param offset starting index in a list of results
    * @param limit  limit of records for pagination
+   * @param tenantId tenant id
    * @return future with {@link SourceRecordCollection}
    */
-  Future<SourceRecordCollection> getSourceRecords(String query, int offset, int limit);
+  Future<SourceRecordCollection> getSourceRecords(String query, int offset, int limit, String tenantId);
 }